Adding The Date of The Leaked Database in The Excel File
DEHASHED has a tab which is called 'Data Wells' where all the databases
that the service is checking for leaked credentials and other information reside.
Since I didn't found a way to automatically pull the data, I copy paste it to a
XLSX file which make it succeptible to changes on the service's data well.

Moreover, additional code was added in the __convert_json_data_to_csv private method
which takes the parsed json introduces a new dict key, 'date' with a default value of
'No Date'. In case there is a much in the data well that value is replace by the date
that the database was leaked.

Finally, another private method was added under the name of __read_dehashed_datawell(self).
This method is responsible for reading the dehashed data well from the excel file.
It returns a dictionary with the name of the leaked database as a key and the date the
database was leaked as its value.